getting video data
